+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CICD: Update release workflow images to `ubuntu-latest`
 - Spine: Fix logging when substepping a simulation spine (bis repetita placent)
 
+### Removed
+
+- envs: Remove UpkieServoPositions and UpkieServoTorques environments
+
 ## [8.0.0] - 2025-05-08
 
 #### Added